# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Provincias - TP8 DAI",
    "description": "Documentación generada con Swagger para la API de provincias del TP8. Permite consultar, crear, actualizar y eliminar provincias.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Provincias",
      "description": "Operaciones sobre el recurso Provincia"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/province/": {
      "get": {
        "tags": [
          "Provincias"
        ],
        "summary": "Obtiene todas las provincias",
        "description": "Devuelve el listado completo de provincias almacenadas en la base de datos.",
        "responses": {
          "200": {
            "description": "Listado de provincias obtenido correctamente",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Province"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Provincias"
        ],
        "summary": "Crea una nueva provincia",
        "description": "Da de alta una provincia nueva. El id se genera automáticamente en la base de datos.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Datos de la provincia a crear",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProvinceInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Provincia creada correctamente"
          },
          "400": {
            "description": "Datos inválidos (por ejemplo, falta el nombre o el full_name)"
          },
          "500": {
            "description": "Error interno al intentar crear la provincia"
          }
        }
      },
      "put": {
        "tags": [
          "Provincias"
        ],
        "summary": "Actualiza una provincia existente",
        "description": "Modifica los datos de una provincia ya existente. El id debe venir incluido en el body.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Datos de la provincia a actualizar, incluyendo su id",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Province"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Provincia actualizada correctamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "404": {
            "description": "No existe una provincia con ese id"
          },
          "500": {
            "description": "Error interno al intentar actualizar la provincia"
          }
        }
      }
    },
    "/api/province/{id}": {
      "get": {
        "tags": [
          "Provincias"
        ],
        "summary": "Obtiene una provincia por id",
        "description": "Devuelve los datos de una provincia puntual a partir de su id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Id de la provincia"
          }
        ],
        "responses": {
          "200": {
            "description": "Provincia encontrada",
            "schema": {
              "$ref": "#/definitions/Province"
            }
          },
          "404": {
            "description": "No existe una provincia con ese id"
          }
        }
      },
      "delete": {
        "tags": [
          "Provincias"
        ],
        "summary": "Elimina una provincia",
        "description": "Elimina la provincia cuyo id coincide con el parámetro recibido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Id de la provincia a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Provincia eliminada correctamente"
          },
          "404": {
            "description": "No existe una provincia con ese id"
          }
        }
      }
    }
  },
  "definitions": {
    "Province": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Buenos Aires"
        },
        "full_name": {
          "type": "string",
          "example": "Provincia de Buenos Aires"
        },
        "latitude": {
          "type": "number",
          "example": -34.6
        },
        "longitude": {
          "type": "number",
          "example": -58.4
        },
        "display_order": {
          "type": "number",
          "example": 1
        }
      }
    },
    "ProvinceInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Buenos Aires"
        },
        "full_name": {
          "type": "string",
          "example": "Provincia de Buenos Aires"
        },
        "latitude": {
          "type": "number",
          "example": -34.6
        },
        "longitude": {
          "type": "number",
          "example": -58.4
        },
        "display_order": {
          "type": "number",
          "example": 1
        }
      }
    }
  }
}